Update nearestYard.py
type casting to float

+++ Copart/yard problem/nearestYard.py
@@ -30,8 +30,9 @@ def distance(lat1, lng1, lat2, lng2):
     return radius * ang
 #author vinaya
 array = []
-lat2 = sys.argv[1]
-lng2 = sys.argv[2]
+lat2 = (float)sys.argv[1]
+lng2 = (float)sys.argv[2]
+
 
 
 for index, row in dataset.iterrows():
